Fix test case for ARM

+++ tests/output/c-backend/cast-ptr-arithm-arm.c
@@ -1,4 +1,4 @@
 void *func(void *p1, char *p2, char *p3)
 {
-  return ( void *) ( ( int ) ( p2 + ( unsigned int ) p1 ) - ( int ) p3 );
+  return ( void *) ( p2 + ( unsigned int ) p1 - p3 );
 }
